Update My_Useful_Fns.R
This is an aesthetic commit, mainly. I do switch from cbind to merge to better guard against errors caused by dropped genes in make_bigwig_score_matrix, and I also fix an off-by-one error that shows up in both that function and the bam score matrix one, because bed is 0-based but G-ranges is 1-based.